chore: extract parent TestingHelper class
